Add authorization checks for delete operations in exams, questions, and subjects

+++ backend/controllers/exams.js
@@ -2,8 +2,10 @@ const ExamModel = require("../database/models/exam");
 const ExamQuestionModel = require("../database/models/exam_question");
 const CustomError = require("../utils/customError");
 const { validateExam } = require("../utils/validator");
-const { BAD_REQUEST, SUCCESS } = require("../constants/statusCodes");
+const { BAD_REQUEST, SUCCESS, FORBIDDEN } = require("../constants/statusCodes");
+const UserModel = require("../database/models/user");
 
+const User = new UserModel();
 const Exam = new ExamModel();
 const ExamQuestion = new ExamQuestionModel();
 
@@ -50,8 +52,18 @@ const editExam = async (req, res) => {
 const deleteExam = async (req, res) => {
   const userId = res.locals.userId;
   const examId = req.params.id;
-  console.log(userId, examId);
+  const user = await User.getUserInfo(userId);
+  const whitelist = ["Admin", "Faculty"];
+
+  if (!whitelist.includes(user?.role)) {
+    throw new CustomError(
+      "You are not authorized to perform this action",
+      FORBIDDEN
+    );
+  }
+
   const result = await Exam.deleteExam(userId, examId);
+  console.log(result);
   res.send({ message: "Successfully deleted.", result });
 };
